Code for the Add/View  Water Quality Table
Nhia,
Adding code for the add/view data for the water quality table. I have
tested with my local server. just not tested with the  UGA staging
server. Assuming if it works on my local it does with the uga server.
There is dbHelpers.dao file that has the details for the connection
string, you will need to change that to test against the UGA staging
environment.

Everyone, when writing your code please import the dbHelpers.dao class
so it becomes easy to form the connection and you do not need to create
one for every single action. you can also copy my code for that is
needed.

 Brian will need you to apply some magic with the  styling on the
ViewData.jsp
I am working on the add/View data for the monitoring Table and should
have that by this weekend.